+++ src/sponge/mod.rs
@@ -4,6 +4,8 @@
 //
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
-pub mod sponge;
+mod sponge;
 
+#[cfg(feature = "std")]
+pub use sponge::sponge_gadget as gadget;
 pub use sponge::sponge_hash as hash;